Add automatic update checking feature
A new class UpdateChecker has been added to the TrayApp that periodically checks for new releases using an HTTP client. The current version of the application is compared against the latest version fetched from a specific URL. If a new update is available, a message box is shown to the user prompting them to download the latest version from GitHub. Error handling has been implemented to notify the user if there's any failure during the process. The feature is integrated into the program startup sequence. This addition will help ensure users always have the most up-to-date version of the software.